# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BE-01 Task API",
    "version": "Stage 6.4"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "Tasks",
      "description": "CRUD operations on the in-memory task store."
    },
    {
      "name": "Meta",
      "description": "Service metadata and health endpoints."
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Returns API metadata and the list of routes.",
        "responses": {
          "200": {
            "description": "API name, version and available endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "version",
                    "endpoints"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Task API"
                    },
                    "version": {
                      "type": "string",
                      "example": "Stage 6.3"
                    },
                    "endpoints": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "/docs",
                        "/tasks",
                        "/health",
                        "/stats",
                        "/tasks/{id}",
                        "POST /tasks",
                        "POST /reset",
                        "PUT /tasks/{id}",
                        "DELETE /tasks/{id}"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Checks whether the server is up.",
        "responses": {
          "200": {
            "description": "Server is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Returns aggregate counts of the task store.",
        "description": "Totals derived from the in-memory store: `total` is the number of tasks, `done` counts completed tasks, and `open` is `total - done`.",
        "responses": {
          "200": {
            "description": "Task counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "done",
                    "open"
                  ],
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "Total number of tasks.",
                      "example": 3
                    },
                    "done": {
                      "type": "integer",
                      "description": "Number of completed tasks.",
                      "example": 1
                    },
                    "open": {
                      "type": "integer",
                      "description": "Number of incomplete tasks (`total - done`).",
                      "example": 2
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Returns a list of tasks, optionally filtered by completion state and/or a title substring.",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "description": "Filter by completion state. Omit to return all tasks.",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "example": "true"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive substring match against task titles. Omit to return all tasks.",
            "schema": {
              "type": "string",
              "example": "sample"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of matching tasks (all tasks when no filter is supplied).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "400": {
            "description": "A query parameter was supplied with an invalid value: `done` must be true or false, and `search` must be a string.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidDone": {
                    "value": {
                      "error": "Invalid done query"
                    }
                  },
                  "invalidSearch": {
                    "value": {
                      "error": "Invalid search query"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Creates a new task from a title.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Payload for creating a task. `done` is rejected if present.",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Non-empty task title.",
                    "example": "Write the OpenAPI spec"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created; done always starts as false.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "done was provided, or title is missing/not a non-empty string.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "doneNotSupported": {
                    "value": {
                      "error": "done is not supported on POST"
                    }
                  },
                  "invalidTitle": {
                    "value": {
                      "error": "Missing or invalid title"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Content-Type is not application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Expected application/json"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Integer task ID (must be non-zero).",
          "schema": {
            "type": "integer",
            "format": "int32",
            "example": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Returns a specific task by ID.",
        "responses": {
          "200": {
            "description": "The requested task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "ID is not an integer or is zero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid ID"
                }
              }
            }
          },
          "404": {
            "description": "No task exists with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 42 not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Updates an existing task; accepts partial updates.",
        "description": "Supports partial updates: omit `title` or `done` to leave that field unchanged.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Partial update payload; at least one field is expected.",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "New non-empty task title.",
                    "example": "Write the OpenAPI spec"
                  },
                  "done": {
                    "type": "boolean",
                    "description": "New completion state.",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "ID is invalid, or a provided title/done value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidId": {
                    "value": {
                      "error": "Invalid ID"
                    }
                  },
                  "invalidTitle": {
                    "value": {
                      "error": "Invalid title"
                    }
                  },
                  "invalidDone": {
                    "value": {
                      "error": "Invalid done"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No task exists with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 42 not found"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type is not application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Expected application/json"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Deletes an existing task by ID.",
        "responses": {
          "204": {
            "description": "Task deleted; no content returned."
          },
          "400": {
            "description": "ID is not an integer or is zero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid ID"
                }
              }
            }
          },
          "404": {
            "description": "No task exists with that ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Task 42 not found"
                }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Resets the task store to the seed data.",
        "description": "Clears the in-memory store and reloads the tasks from `seed.json`, restoring the initial tasks and the next-ID counter.",
        "responses": {
          "200": {
            "description": "The store was reset; returns the restored tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "description": "A single task.",
        "required": [
          "id",
          "title",
          "done"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Sample Task A"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Error": {
        "type": "object",
        "description": "Error envelope returned by all failure responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Task 42 not found"
          }
        }
      }
    }
  }
}